drivers: Work on stm32f4cube_eth
Extract common code from stm32cube_eth
