+++ src/drivers/net/stm32cube/stm32f4cube_eth.h
@@ -1,0 +1,15 @@
+/**
+ * @file
+ *
+ * @date 17.03.2017
+ * @author Anton Bondarev
+ */
+
+#ifndef STM32F4CUBE_ETH_H_
+#define STM32F4CUBE_ETH_H_
+
+#include <stm32f4xx.h>
+#include <stm32f4xx_hal.h>
+#include <stm32f4xx_hal_eth.h>
+
+#endif /* STM32F4CUBE_ETH_H_ */
